+++ qcetl/api.py
@@ -43,6 +43,7 @@ import qcetl.runscanner.illumina
 import qcetl.samtools
 import qcetl.sequenza
 import qcetl.picard.umiconsensus
+import qcetl.ultimamethylcontrol
 import qcetl.ultimalibrarymetrics
 import qcetl.umiqc
 import qcetl.xenoclassify
@@ -103,6 +104,7 @@ formats = (
     qcetl.samtools.SamtoolsStatsCov2Cache(),
     qcetl.sequenza.SequenzaCache(),
     qcetl.picard.umiconsensus.HsMetricsUmiConsensusCache(),
+    qcetl.ultimamethylcontrol.UltimaMethylControlCache(),
     qcetl.ultimalibrarymetrics.UltimaLibraryMetricsCache(),
     qcetl.umiqc.umiQcCache(),
     qcetl.xenoclassify.XenoclassifyCache(),