+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * ultimalibrarymetrics input records now require a `barcode`, and the cache is keyed by run and barcode name
 
+### Added
+* ultimamethylcontrol cache
+
 ## [1.48] - 2026-09-02
 ### Added
 * ultimalibrarymetrics cache